% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_hero_players.R
\name{get_hero_players}
\alias{get_hero_players}
\title{Dota Hero Players}
\usage{
get_hero_players(patch = "7.19", split = "lan", after = "2011.01.01",
  before = "2018.07.12", duration = "0-100", leagues = NULL,
  tier = "1&2&3", valve = "does-not-matter")
}
\arguments{
\item{patch}{Which patches are you interested in? From 6.74 to 7.20. (Connected by "&")}

\item{split}{What type of event was it? Choose from "online", "lan" or "post-event".}

\item{after}{The beginning date of the match}

\item{before}{The last day of the period to search}

\item{duration}{The time the game lasts (measured in minutes)}

\item{leagues}{What league is it in? Check the serial number of the league on www.datdota.com.}

\item{tier}{What tier of event is this? 1 = Premium; 2 = Professional; 3 = Semi-pro; (Connected by "&")}

\item{valve}{Is this a Valve event? Choose from "yes", "no" or "does-not-matter".}
}
\value{
A dota_api object containing the elements described in the details.
}
\description{
A list of the most frequent players of particular heroes.
}
\details{
A list will be returned that contains three elements. The content, the url and the response
received from the api.

The content element of the list contains a list includes hero, games, rank and players.

\itemize{
  \item \strong{hero:} The serial number of the hero.
  \item \strong{games:} Number of use.
  \item \strong{rank:} Rank among users. Includes 1, 2 and 3.
  \item \strong{players:} A list contains players.
  \item \strong{steam ID:} The steam ID of the player.
  \item \strong{nickname:} The nickname of the player.
}
}
\examples{
\dontrun{
get_heroes_players()
get_heroes_players(patch = "7.19", split = "lan", after = "2011.01.01", before = "2018.07.12",
                   duration = "0-100", leagues = NULL, tier = "1&2&3", valve = "does-not-matter")
}

}
