% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_match_length.R
\name{get_match_length}
\alias{get_match_length}
\title{Match Length Distributions}
\usage{
get_match_length(patch = "7.19", split = "lan", after = "2011.01.01",
  before = "2018.07.12", duration = "0-100", leagues = NULL,
  tier = "1&2&3", valve = "does-not-matter", teams = NULL)
}
\arguments{
\item{patch}{Which patches are you interested in? From 6.74 to 7.20. (Connected by "&")}

\item{split}{What type of event was it? Choose from "online", "lan" or "post-event".}

\item{after}{The beginning date of the match}

\item{before}{The last day of the period to search}

\item{duration}{The time the game lasts (measured in minutes)}

\item{leagues}{What league is it in? Check the serial number of the league on www.datdota.com.}

\item{tier}{What tier of event is this? 1 = Premium; 2 = Professional; 3 = Semi-pro; (Connected by "&")}

\item{valve}{Is this a Valve event? Choose from "yes", "no" or "does-not-matter".}

\item{teams}{Teams to serch for. Check the serial number of the players on www.datdota.com. (Connected by "&". At least one must appear)}
}
\value{
A dota_api object containing the elements described in the details.
}
\description{
Three lists of the duration of matches
}
\details{
A list will be returned that contains three elements. The content, the url and the response
received from the api.

The content element of the list contains three sub lists. The first list is called durations which
tells the distribution of the minutes. The second list is called longest which includes the match id, teams
and duration. The third list is called shortest which includes the match id, teams and duration. There are
also three variables in the content list which provides the mean, the standard deviation and the count.

\itemize{
  \item \strong{minute:} The time the game lasts.
  \item \strong{percentile:} Percentage.
  \item \strong{match:} The match ID.
  \item \strong{teams:} Two teams of this match.
  \item \strong{duration:} The time the game lasts.
  \item \strong{mean:} Average time of these matches.
  \item \strong{stdDev} Standard deviation.
  \item \strong{count:} Total number of matches.
}
}
\examples{
\dontrun{
get_match_length()
get_match_length(patch = "7.19", split = "lan", after = "2011.01.01", before = "2018.07.12",
                 duration = "0-100", leagues = NULL, tier = "1&2&3", valve = "does-not-matter",
                 teams = "15")
}

}
