% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Core.R
\name{get_response}
\alias{get_response}
\title{The engine of the request functions}
\usage{
get_response(dota_api_category, dota_api_subcategory, args)
}
\arguments{
\item{dota_api_category}{One of heroes, players, teams, matchfinder, match-durations}

\item{dota_api_subcategory}{Child entry of category}

\item{args}{The query arguments for GET.}
}
\value{
A response to be used in the rest of the get_* family functions.
}
\description{
The function used to fetch the response content inside the the API functions.
}
\details{
This function is of no particular use to the user, but might be useful for anyone who
would like to dive deeper into the package development. This function is used to GET the
response's content from Steam's API. The function is the core function of the rest of
the get_* family functions.
}
\examples{
\dontrun{
get_response("heroes","performances", args = list(patch = "7.20"))
}

}
