% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/para_trans.R
\name{para_trans}
\alias{para_trans}
\title{The pre process of the parameters of the request functions}
\usage{
para_trans(after = "2011.01.01", before = "2018.01.01",
  duration = "0-100", tier = "1&2&3", patch = "7.20&7.19",
  split = "online&lan", leagues = "NULL", valve = "does-not-matter")
}
\arguments{
\item{after}{The beginning date of the match}

\item{before}{The last day of the period to search}

\item{duration}{The time the game lasts (measured in minutes)}

\item{tier}{What tier of event is this? 1 = Premium; 2 = Professional; 3 = Semi-pro; (Connected by "&")}

\item{patch}{Which patches are you interested in? From 6.74 to 7.20. (Connected by "&")}

\item{split}{What type of event was it? Choose from "online", "lan" or "post-event".}

\item{leagues}{What league is it in? Check the number of the league on www.datdota.com.}

\item{valve}{Is this a Valve event? Choose from "yes", "no" or "does-not-matter".}
}
\value{
A list of parameters used in get_response function.
}
\description{
This function is of no particular use to the user, but might be useful for anyone who
would like to dive deeper into the package development. This function is used to pre edit
the parameters of the GET functions
}
\examples{
\dontrun{
para_trans()
para_trans(after = "2012.01.01", before = "2018.01.01", duration = "0-50", tier = "1")
}

}
